// frontend/src/types.ts
export interface InvestorData {
    vcName?: string;
    investorName?: string;
    linkedinUrl?: string;
    focusAreas?: string[];
    geographical?: string[];
    targetCountries?: string[];
    fundingRequirements?: string;
    fundingStages?: string[];
    checkSize?: string;
    url?: string;
    description?: string;
}